intialize core data
