+++ Artilces/Artilces/CoreData/Articles+CoreDataClass.swift
@@ -1,0 +1,16 @@
+//
+//  Articles+CoreDataClass.swift
+//  Artilces
+//
+//  Created by Agam Mahajan on 01/03/18.
+//  Copyright © 2018 Agam Mahajan. All rights reserved.
+//
+//
+
+import Foundation
+import CoreData
+
+
+public class Articles: NSManagedObject {
+
+}
